requirements.txt
azure-functions
arrow
pandas
azure-core
azure-storage-file-datalake
azure-identity
pathlib
uri